+++ node/__tests__/unit/middlewares/parseData.spec.ts
@@ -2,7 +2,10 @@ import type { EventContext } from '@vtex/api'
 
 import type { Clients } from '../../../clients'
 import { parseData } from '../../../middlewares/parseData'
-import { CUSTOM_DATA_FIELD_ID } from '../../../utils/constants'
+import {
+  CUSTOM_DATA_FIELD_ID,
+  LOGGER_ERROR_MESSAGES,
+} from '../../../utils/constants'
 
 describe('parseData middleware', () => {
   const next = jest.fn()
@@ -20,7 +23,7 @@ describe('parseData middleware', () => {
           totals: [{ id: 'Items', value: 100 }],
           items: [
             {
-              id: 'skuId',
+              id: '1',
               skuName: 'skuName',
               imageUrl: 'url',
               sellingPrice: 100,
@@ -32,6 +35,11 @@ describe('parseData middleware', () => {
           fields: {
             [CUSTOM_DATA_FIELD_ID]: 'affiliateId',
           },
+        },
+      },
+      clients: {
+        commissionBySKU: {
+          get: jest.fn().mockResolvedValueOnce([{ id: '1', commission: 10 }]),
         },
       },
       vtex: {
@@ -56,10 +64,10 @@ describe('parseData middleware', () => {
             email: 'clientEmail@email.com',
           },
           creationDate: 'date',
-          totals: [],
+          totals: null,
           items: [
             {
-              id: 'skuId',
+              id: '1',
               skuName: 'skuName',
               imageUrl: 'url',
               sellingPrice: 100,
@@ -73,6 +81,11 @@ describe('parseData middleware', () => {
           },
         },
       },
+      clients: {
+        commissionBySKU: {
+          get: jest.fn().mockResolvedValueOnce([{ id: '1', commission: 10 }]),
+        },
+      },
       vtex: {
         logger: {
           error: jest.fn(),
@@ -81,7 +94,7 @@ describe('parseData middleware', () => {
     } as unknown as EventContext<Clients>
 
     return expect(parseData(ctxMock, next)).rejects.toThrow(
-      'Error parsing the order data'
+      LOGGER_ERROR_MESSAGES.parseData
     )
   })
 })