+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Get the variable sku commission in the parseData middleware
+- Update order types
+
 ## [0.6.0] - 2022-01-12
 
 ## Added